refactored and reused a single helper method
